@@ -12,23 +12,23 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "2.4.0",
-    "@angular/compiler": "2.4.0",
-    "@angular/core": "2.4.0",
-    "@angular/forms": "2.4.0",
-    "@angular/http": "2.4.0",
+    "@angular/common": "4.3.1",
+    "@angular/compiler": "4.3.1",
+    "@angular/core": "4.3.1",
+    "@angular/forms": "4.3.1",
+    "@angular/http": "4.3.1",
     "@angular/material": "2.0.0-beta.2",
-    "@angular/platform-browser": "2.4.8",
-    "@angular/platform-browser-dynamic": "2.4.0",
-    "@angular/router": "3.4.0",
+    "@angular/platform-browser": "4.3.1",
+    "@angular/platform-browser-dynamic": "4.3.1",
+    "@angular/router": "4.3.1",
     "core-js": "2.4.1",
     "hammerjs": "2.0.8",
-    "rxjs": "5.1.0",
-    "zone.js": "0.7.6"
+    "rxjs": "5.4.2",
+    "zone.js": "0.8.5"
   },
   "devDependencies": {
-    "@angular/cli": "1.0.0-beta.32.3",
-    "@angular/compiler-cli": "2.4.0",
+    "@angular/cli": "^1.2.4",
+    "@angular/compiler-cli": "4.3.1",
     "@types/hammerjs": "2.0.34",
     "@types/jasmine": "2.5.38",
     "@types/node": "~6.0.60",
@@ -44,6 +44,6 @@
     "protractor": "~5.1.0",
     "ts-node": "~2.0.0",
     "tslint": "~4.4.2",
-    "typescript": "~2.0.0"
+    "typescript": "^2.4.2"
   }
 }
