@@ -5,7 +5,7 @@ import { Http, Response, Request, Headers, RequestOptions } from '@angular/http'
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/catch';
 import 'rxjs/add/observable/throw';
-import { Observable } from "rxjs/Observable";
+import { Observable } from 'rxjs/Observable';
 import { HttpServices } from 'app/AppCommon/httpservices.service';
 
 @Injectable()
@@ -17,17 +17,18 @@ export class CallWaitingService {
     getCallWaitingService(acrUrl, postCWGet) {
         this.httpservices.httpGetRequest(acrUrl)
             .subscribe((res) => {
-                var cwsParsedJson = res.json();
+                let cwsParsedJson = res.json();
                 postCWGet(cwsParsedJson);
             }, (err) => {
-                console.log("Some Error in CWS.");
+                console.log('Some Error in CWS.');
                 postCWGet(null);
             });
     }
 
     putCallWaitingService(cwsUrl, isCallWaitingChecked, postCWSPut) {
 
-        var body = '<?xml version="1.0" encoding="UTF-8"?><CallWaiting xmlns="http://schema.broadsoft.com/xsi"><active>' + isCallWaitingChecked + '</active></CallWaiting>';
+        let body = '<?xml version="1.0" encoding="UTF-8"?><CallWaiting xmlns="http://schema.broadsoft.com/xsi"><active>'
+                    + isCallWaitingChecked + '</active></CallWaiting>';
         this.httpservices.httpPutRequest(cwsUrl, body)
             .subscribe((res) => {
                 postCWSPut(res)
