@@ -4,7 +4,7 @@ import { Injectable, Inject, Component } from '@angular/core';
 import { Http, Response, Request, Headers, RequestOptions } from '@angular/http';
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/catch';
-import { Observable } from "rxjs/Observable";
+import { Observable } from 'rxjs/Observable';
 import 'rxjs/add/observable/throw';
 import { HttpServices } from 'app/AppCommon/httpservices.service';
 
@@ -18,18 +18,19 @@ export class DoNotDisturbService {
 
         this.httpServices.httpGetRequest(dndUrl)
             .subscribe((res) => {
-                var dndParsedJson = res.json();
+                let dndParsedJson = res.json();
                 postDndGet(dndParsedJson);
 
             }, (err) => {
                 console.log('Error Occurred!!!')
-                var dndParsedJson = null;
+                let dndParsedJson = null;
                 postDndGet(dndParsedJson);
             });
     }
 
     putDoNotDisturbService(dndUrl, isDndChecked, postDndPut) {
-        var body = '<?xml version="1.0" encoding="UTF-8"?><DoNotDisturb xmlns="http://schema.broadsoft.com/xsi"><active>' + isDndChecked + '</active><ringSplash>false</ringSplash></DoNotDisturb>';
+        let body = '<?xml version="1.0" encoding="UTF-8"?><DoNotDisturb xmlns="http://schema.broadsoft.com/xsi"><active>'
+                    + isDndChecked + '</active><ringSplash>false</ringSplash></DoNotDisturb>';
         this.httpServices.httpPutRequest(dndUrl, body)
             .subscribe((res) => {
                 postDndPut(res);
