@@ -1,7 +1,7 @@
 /* Copyright © 2017 BroadSoft Inc. */
 
 import { Injectable, Inject, Component } from '@angular/core';
-import { Observable } from "rxjs/Observable";
+import { Observable } from 'rxjs/Observable';
 import { HttpServices } from 'app/AppCommon/httpservices.service';
 
 @Injectable()
@@ -13,7 +13,7 @@ export class AutomaticCallbackService {
 
         this.httpservices.httpGetRequest(acbUrl)
             .subscribe((res) => {
-                var acbParsedJson = res.json();
+                let acbParsedJson = res.json();
                 postACBGet(acbParsedJson);
             }, (err) => {
                 postACBGet(null);
@@ -22,7 +22,9 @@ export class AutomaticCallbackService {
 
     putAutomaticCallbackService(acbUrl, isAutomaticCallbackChecked, postACBPut) {
 
-        var body = '<?xml version="1.0" encoding="UTF-8"?> <AutomaticCallback xmlns="http://schema.broadsoft.com/xsi"><active>' + isAutomaticCallbackChecked + '</active></AutomaticCallback>';
+        let body = '<?xml version="1.0" encoding="UTF-8"?> <AutomaticCallback xmlns="http://schema.broadsoft.com/xsi"><active>'
+                    + isAutomaticCallbackChecked + '</active></AutomaticCallback>';
+
         this.httpservices.httpPutRequest(acbUrl, body)
             .subscribe((res) => {
                 postACBPut(res);
