@@ -1,59 +1,72 @@
 /* Copyright © 2017 BroadSoft Inc. */
 
-import { Component, OnInit } from '@angular/core';
-import { XSIServices } from 'app/AppCommon/xsiServiceList.service';
-import { HttpServices } from 'app/AppCommon/httpservices.service';
-import { ServiceRouteProvider } from 'app/AppCommon/serviceRouteProvider.service';
-import { ServiceRouter } from 'app/AppCommon/serviceRouter.service';
-import { RemoteOfficeServiceInput } from 'app/IncomingCalls/RemoteOffice/remoteOfficeServiceInput.service';
-import { XSIMasterServicesList } from 'app/AppCommon/xsiMasterServicesList.service';
-import { MasterServiceListVisible } from 'app/AppCommon/masterServicesListVisible.service';
-import { CallForwardingService } from 'app/IncomingCalls/CallForwarding/callForwardingService.service';
-import { CallForwardingServiceInput } from 'app/IncomingCalls/CallForwarding/callForwardingServiceInput.service';
-import { CallControlServiceInput } from 'app/CallControl/callControlServiceInput.service';
-import { BroadWorksMobilityServiceInput } from 'app/CallControl/BroadWorksMobility/broadWorksMobilityServiceInput.service';
-import { CallControlService } from 'app/CallControl/callControlService.service'
-import { SimultaneousRingService } from 'app/IncomingCalls/SimultaneousRing/simultaneousRing.service';
-import { SimultaneousRingServiceInput } from 'app/IncomingCalls/SimultaneousRing/simultaneousRingServiceInput.service';
-import { VoicemailService } from 'app/Voicemail/voicemailService.service';
-import { VoicemailServiceInput } from 'app/Voicemail/voicemailServiceInput.service';
+import {Component, OnInit} from '@angular/core';
+import {XSIServices} from 'app/AppCommon/xsiServiceList.service';
+import {HttpServices} from 'app/AppCommon/httpservices.service';
+import {ServiceRouteProvider} from 'app/AppCommon/serviceRouteProvider.service';
+import {ServiceRouter} from 'app/AppCommon/serviceRouter.service';
+import {RemoteOfficeServiceInput} from 'app/IncomingCalls/RemoteOffice/remoteOfficeServiceInput.service';
+import {XSIMasterServicesList} from 'app/AppCommon/xsiMasterServicesList.service';
+import {MasterServiceListVisible} from 'app/AppCommon/masterServicesListVisible.service';
+import {CallForwardingService} from 'app/IncomingCalls/CallForwarding/callForwardingService.service';
+import {CallForwardingServiceInput} from 'app/IncomingCalls/CallForwarding/callForwardingServiceInput.service';
+import {BroadWorksMobilityServiceInput} from 'app/CallControl/BroadWorksMobility/broadWorksMobilityServiceInput.service';
+import {CallControlServiceInput} from 'app/CallControl/callControlServiceInput.service';
+import {SimultaneousRingService} from 'app/IncomingCalls/SimultaneousRing/simultaneousRing.service';
+import {SimultaneousRingServiceInput} from 'app/IncomingCalls/SimultaneousRing/simultaneousRingServiceInput.service';
+import {VoicemailService} from 'app/Voicemail/voicemailService.service';
+import {VoicemailServiceInput} from 'app/Voicemail/voicemailServiceInput.service';
+import {SequentialRingServiceInput} from 'app/IncomingCalls/SequentialRing/sequentialRingServiceInput.service';
+import {SequentialRingService} from 'app/IncomingCalls/SequentialRing/sequentialRing.service';
+import {CallControlService} from 'app/CallControl/callControlService.service';
+import {CallCenterService} from 'app/CallControl/CallCenterQueues/callcenterService.service';
+import {CallCenterServiceInput} from 'app/CallControl/CallCenterQueues/callcenterServiceInput.service';
+import {OutOfOfficeServiceInput} from 'app/IncomingCalls/OutOfOffice/outOfOfficeServiceInput.service';
+import {BroadworksAnywhereService} from 'app/CallControl/BroadworksAnywhere/broadworksAnywhereService.service';
+import {BroadworksAnywhereServiceInput} from 'app/CallControl/BroadworksAnywhere/broadworksAnywhereServiceInput.service';
+import {BroadWorksAnywhereLocationsInput} from 'app/CallControl/BroadworksAnywhere/broadworksAnywhereLocationsInput.service';
+import {Util} from 'app/AppCommon/util';
 
 @Component({
   selector: 'bsft-call-settings',
   templateUrl: 'app.component.html',
-  providers: [BroadWorksMobilityServiceInput, CallControlServiceInput, CallForwardingServiceInput, CallForwardingService,
-    MasterServiceListVisible, XSIMasterServicesList, XSIServices, HttpServices, ServiceRouteProvider,
-    ServiceRouter, RemoteOfficeServiceInput, SimultaneousRingService, SimultaneousRingServiceInput,
-    VoicemailService, VoicemailServiceInput, CallControlService]
+  providers: [BroadWorksMobilityServiceInput, CallControlServiceInput, CallForwardingServiceInput,
+    CallForwardingService, MasterServiceListVisible, XSIMasterServicesList, XSIServices, HttpServices,
+    ServiceRouteProvider, ServiceRouter, RemoteOfficeServiceInput, SimultaneousRingService,
+    SimultaneousRingServiceInput, VoicemailService, VoicemailServiceInput, CallControlService,
+    CallCenterService, CallCenterServiceInput, OutOfOfficeServiceInput, BroadworksAnywhereService,
+    BroadworksAnywhereServiceInput, BroadWorksAnywhereLocationsInput, SequentialRingServiceInput,
+    Util]
 })
 
 /*This class handles the UI components for the application*/
-export class AppComponent {
+export class AppComponent implements OnInit {
 
   customizedTextJson = window['customizedTexts'];
 
   xsiServicesList: string;
-  isIncomingCallExpanded: boolean = false;
-  isOutgoingCallExpanded: boolean = false;
-  isVoiceMailExpanded: boolean = false;
-  isCallControlExpanded: boolean = false;
+  isIncomingCallExpanded = false;
+  isOutgoingCallExpanded = false;
+  isVoiceMailExpanded = false;
+  isCallControlExpanded = false;
   masterServicesList = new Array();
-  private isReady: boolean = false;
+  isReady = false;
   isIncomingCallsVisible: boolean;
   isOutgoingCallsVisible: boolean;
   isCallControlVisible: boolean;
   isVoiceMailVisible: boolean;
   isNoServiceVisible: boolean;
 
-  constructor(private xsiMasterServicesList: XSIMasterServicesList, private serviceRouteProvider: ServiceRouteProvider, private xsiServices: XSIServices) {
+  constructor(private xsiMasterServicesList: XSIMasterServicesList, private serviceRouteProvider: ServiceRouteProvider,
+    private xsiServices: XSIServices) {
 
   }
 
   ngOnInit() {
 
-    /*Generates the login token for dev/QA purposes only*/
-    console.log("Application mode : " + window['applicationMode']);
-    if (window['applicationMode'] == 'dev' || window['applicationMode'] == 'qa') {
+      /*Generates the login token for dev/QA purposes only*/
+    console.log('Application mode : ' + window['applicationMode']);
+    if (window['applicationMode'] === 'dev' || window['applicationMode'] === 'qa') {
       this.xsiServices.generateLoginTokenAndUse(this.postGeneratedLoginToken.bind(this));
     } else {
       if (window['isTextsLoaded']) {
@@ -75,7 +88,8 @@ export class AppComponent {
       /*No services is to be shown */
       console.log("No services is to be shown ");
     } else { /*Check the visibility criteria for the section headers*/
-      console.log('Master Services List: ', this.xsiServices.fetchMasterServicesList());     
+      console.log('Master Services List: ', this.xsiServices.fetchMasterServicesList());
+
       this.checkOutgoingCallsVisible();
       this.checkIncomingCallsVisible();
       this.checkVoiceMailVisible();
@@ -95,8 +109,8 @@ export class AppComponent {
 
   checkOutgoingCallsVisible() {
 
-    var bcidVisible: boolean = false;
-    var acbVisible: boolean = false;
+    let bcidVisible = false;
+    let acbVisible = false;
     if (this.serviceRouteProvider.fetchBlockCallerIdUrl()) {
       bcidVisible = this.xsiServices.fetchBcidVisible();
     } else {
@@ -116,15 +130,16 @@ export class AppComponent {
 
   checkIncomingCallsVisible() {
 
-    var isDNDVisible: boolean = false;
-    var isACRVisible: boolean = false;
-    var isCFAVisible: boolean = false;
-    var isCFBVisible: boolean = false;
-    var isCFNAVisible: boolean = false;
-    var isCFNRVisible: boolean = false;
-    var isSimultaneousRingvisible: boolean = false;
-    var isRemoteOfficeVisible: boolean = false;
-    var isCWVisible: boolean = false;
+    let isDNDVisible = false;
+    let isACRVisible = false;
+    let isCFAVisible = false;
+    let isCFBVisible = false;
+    let isCFNAVisible = false;
+    let isCFNRVisible = false;
+    let isSimultaneousRingvisible = false;
+    let isRemoteOfficeVisible = false;
+    let isCWVisible = false;
+    let isPAVisible = false;
 
     if (this.serviceRouteProvider.fetchDoNotDisturbUrl()) {
       isDNDVisible = this.xsiServices.fetchDndVisible();
@@ -171,8 +186,15 @@ export class AppComponent {
     } else {
       this.xsiServices.setCwVisible(isCWVisible);
     }
+    if (this.serviceRouteProvider.fetchPersonalAssistantUrl()) {
+      isPAVisible = this.xsiServices.fetchPAVisible();
+    }
 
-    if (!isDNDVisible && !isACRVisible && !isCFAVisible && !isCFBVisible && !isCFNAVisible && !isCFNRVisible && !isSimultaneousRingvisible && !isRemoteOfficeVisible && !isCWVisible) {
+    if (!isDNDVisible && !isACRVisible && !isCFAVisible &&
+      !isCFBVisible && !isCFNAVisible && !isCFNRVisible &&
+      !isSimultaneousRingvisible && !isRemoteOfficeVisible
+      && !isCWVisible && !isPAVisible) {
+
       this.isIncomingCallsVisible = false;
     } else {
       this.isIncomingCallsVisible = true;
@@ -181,7 +203,7 @@ export class AppComponent {
 
   checkVoiceMailVisible() {
 
-    var isVoiceMailVisible: boolean = false;
+    let isVoiceMailVisible = false;
     if (this.serviceRouteProvider.fetchVoicemailUrl()) {
       isVoiceMailVisible = this.xsiServices.fetchVoicemailVisible();
     } else {
@@ -196,13 +218,26 @@ export class AppComponent {
 
   checkCallControlVisible() {
 
-    var isBroadWorksMobilityVisible: boolean = false;
+    let isBroadWorksMobilityVisible = false;
+    let isCallCenterVisible = false;
+    let isBroadWorksAnywhereVisible = false;
+
     if (this.serviceRouteProvider.fetchBroadWorksMobilityUrl()) {
       isBroadWorksMobilityVisible = this.xsiServices.fetchBWMobilityVisible();
     } else {
       this.xsiServices.setBWMobilityVisible(isBroadWorksMobilityVisible);
     }
-    if (!isBroadWorksMobilityVisible) {
+    if (this.serviceRouteProvider.fetchCallCenterUrl()) {
+      isCallCenterVisible = this.xsiServices.fetchCallcenterQueueVisible();
+    } else {
+      this.xsiServices.setCallcenterQueueVisible(isCallCenterVisible);
+    }
+    if (this.serviceRouteProvider.fetchBroadWorksAnywhereUrl()) {
+      isBroadWorksAnywhereVisible = this.xsiServices.fetchBroadWorksAnywhereVisible();
+    } else {
+      this.xsiServices.setBroadWorksAnywhereVisible(isBroadWorksAnywhereVisible);
+    }
+    if (!isBroadWorksMobilityVisible && !isCallCenterVisible && !isBroadWorksAnywhereVisible) {
       this.isCallControlVisible = false;
     } else {
       this.isCallControlVisible = true;
@@ -210,30 +245,35 @@ export class AppComponent {
   }
 
   incomingCallExpanded() {
-    if (this.isIncomingCallExpanded)
+    if (this.isIncomingCallExpanded) {
       this.isIncomingCallExpanded = false;
-    else {
+    } else {
       this.isIncomingCallExpanded = true;
-      if (this.isOutgoingCallExpanded || this.isCallControlExpanded)
+      if (this.isOutgoingCallExpanded || this.isCallControlExpanded) {
         this.isOutgoingCallExpanded = false;
-      if (this.isVoiceMailExpanded)
+      }
+      if (this.isVoiceMailExpanded) {
         this.isVoiceMailExpanded = false;
+      }
       this.isCallControlExpanded = false;
     }
   }
 
   outgoingCallExpanded() {
-    if (this.isOutgoingCallExpanded)
+    if (this.isOutgoingCallExpanded) {
       this.isOutgoingCallExpanded = false;
-    else {
+    } else {
       this.isOutgoingCallExpanded = true;
-      if (this.isVoiceMailExpanded)
-        this.isVoiceMailExpanded = false;
-      if (this.isIncomingCallExpanded || this.callControlExpanded)
-        this.isIncomingCallExpanded = false;
-      this.isCallControlExpanded = false;
     }
+    if (this.isVoiceMailExpanded) {
+      this.isVoiceMailExpanded = false;
+    }
+    if (this.isIncomingCallExpanded || this.callControlExpanded) {
+      this.isIncomingCallExpanded = false;
+    }
+    this.isCallControlExpanded = false;
   }
+
 
   callControlExpanded() {
 
@@ -247,21 +287,26 @@ export class AppComponent {
 
       }
     }
-    if (this.isVoiceMailExpanded)
+    if (this.isVoiceMailExpanded) {
       this.isVoiceMailExpanded = false;
+    }
   }
 
   voicemailExpanded() {
-    if (this.isVoiceMailExpanded)
+    if (this.isVoiceMailExpanded) {
       this.isVoiceMailExpanded = false;
-    else
+    } else {
       this.isVoiceMailExpanded = true;
-    if (this.isIncomingCallExpanded)
+    }
+    if (this.isIncomingCallExpanded) {
       this.isIncomingCallExpanded = false;
-    if (this.isOutgoingCallExpanded)
+    }
+    if (this.isOutgoingCallExpanded) {
       this.isOutgoingCallExpanded = false;
-    if (this.isCallControlExpanded)
+    }
+    if (this.isCallControlExpanded) {
       this.isCallControlExpanded = false;
+    }
   }
 
 
