@@ -7,9 +7,10 @@ import { Injectable } from '@angular/core';
 
 /*This class initializes all the service URLs that is to be accessed by the app to fetch and update services' data*/
 export class ServiceRouteProvider {
+    broadWorksAnywhereUrl: string;
 
-    baseUrl: string = window['xsiActionsBaseURL'] + "/v2.0/user/" + window['callSettingsHeroInput'].userId + "/services/";
-    profileUrl: string = window['xsiActionsBaseURL'] + "/v2.0/user/" + window['callSettingsHeroInput'].userId + "/profile/";
+    baseUrl: string = window['xsiActionsBaseURL'] + '/v2.0/user/' + window['callSettingsHeroInput'].userId + '/services/';
+    profileUrl: string = window['xsiActionsBaseURL'] + '/v2.0/user/' + window['callSettingsHeroInput'].userId + '/profile/';
     doNotDisturbUrl: string ;
     anonymousCallRejectionUrl: string ;
     remoteOfficeUrl: string ;
@@ -23,79 +24,102 @@ export class ServiceRouteProvider {
     personalAssistantUrl: string;
     broadWorksMobilityUrl: string;
     simultaneousRingUrl: string;
+    sequentialRingUrl: string;
     voiceMailUrl: string;
     voicemailGreetingUrl: string;
+    callCenterUrl: string;
 
     masterServicesList = [];
 
     constructor(private serviceRouter: ServiceRouter) { }
 
     initializeServiceRoutes(masterServicesList) {
-        
+
         this.masterServicesList = masterServicesList;
         this.masterServicesList.forEach(service => {
 
-            switch(service.name) {
+            switch (service.name) {
 
-                case "Do Not Disturb" : this.doNotDisturbUrl = this.baseUrl + "donotdisturb?";
+                case 'Do Not Disturb' : this.doNotDisturbUrl = this.baseUrl + 'donotdisturb?';
                                         this.serviceRouter.setDoNotDisturbUrl(this.doNotDisturbUrl);
                                         break;
 
-                case "Automatic Callback" : this.automaticCallbackUrl = this.baseUrl + "automaticcallback?";
+                case 'Automatic Callback' : this.automaticCallbackUrl = this.baseUrl + 'automaticcallback?';
                                         this.serviceRouter.setAutomaticCallbackUrl(this.automaticCallbackUrl);
                                         break;
 
-                case "Remote Office" : this.remoteOfficeUrl = this.baseUrl + "remoteoffice?";
+                case 'Remote Office' : this.remoteOfficeUrl = this.baseUrl + 'remoteoffice?';
                                         this.serviceRouter.setRemoteOfficeUrl(this.remoteOfficeUrl);
                                         break;
 
-                case "Call Waiting" : this.callWaitingUrl = this.baseUrl + "callwaiting?";
+                case 'Call Waiting' : this.callWaitingUrl = this.baseUrl + 'callwaiting?';
                                         this.serviceRouter.setCallWaitingUrl(this.callWaitingUrl);
                                         break;
 
-                case "Anonymous Call Rejection" : this.anonymousCallRejectionUrl = this.baseUrl + "anonymouscallrejection?";
+                case 'Anonymous Call Rejection' : this.anonymousCallRejectionUrl = this.baseUrl + 'anonymouscallrejection?';
                                                   this.serviceRouter.setAnonymousCallRejectionUrl(this.anonymousCallRejectionUrl);
                                                   break;
 
-                case "Call Forwarding Always" : this.callForwardingAlwaysUrl = this.baseUrl + "callforwardingalways?";
+                case 'Call Forwarding Always' : this.callForwardingAlwaysUrl = this.baseUrl + 'callforwardingalways?';
                                                 this.serviceRouter.setCallForwardingAlwaysUrl(this.callForwardingAlwaysUrl);
                                                 break;
 
-                case "Call Forwarding Busy" : this.callForwardingBusyUrl = this.baseUrl + "callforwardingbusy?";
+                case 'Call Forwarding Busy' : this.callForwardingBusyUrl = this.baseUrl + 'callforwardingbusy?';
                                               this.serviceRouter.setCallForwardingBusyUrl(this.callForwardingBusyUrl);
                                               break;
 
-                case "Call Forwarding No Answer" : this.callForwardingNoAnswerUrl = this.baseUrl + "callforwardingnoanswer?";
+                case 'Call Forwarding No Answer' : this.callForwardingNoAnswerUrl = this.baseUrl + 'callforwardingnoanswer?';
                                                    this.serviceRouter.setCallForwardingNoAnswerUrl(this.callForwardingNoAnswerUrl);
                                                    break;
 
-                case "Calling Line ID Delivery Blocking" : this.blockCallerIdUrl = this.baseUrl + "callinglineiddeliveryblocking?";
+                case 'Calling Line ID Delivery Blocking' : this.blockCallerIdUrl = this.baseUrl + 'callinglineiddeliveryblocking?';
                                                            this.serviceRouter.setBlockCallerIdUrl(this.blockCallerIdUrl);
                                                            break;
 
-                case "Call Forwarding Not Reachable" : this.callForwardingNotReachableUrl = this.baseUrl + "callforwardingnotreachable?";
-                                                       this.serviceRouter.setCallForwardingNotReachableUrl(this.callForwardingNotReachableUrl);
+                case 'Call Forwarding Not Reachable' : this.callForwardingNotReachableUrl = this.baseUrl + 'callforwardingnotreachable?';
+                                                       this.serviceRouter
+                                                         .setCallForwardingNotReachableUrl(this.callForwardingNotReachableUrl);
                                                        break;
 
-                case "Personal Assistant" : this.personalAssistantUrl = this.baseUrl + "personalassistant?";
+                case 'Personal Assistant' : this.personalAssistantUrl = this.baseUrl + 'personalassistant?';
                                             this.serviceRouter.setPersonalAssistantUrl(this.personalAssistantUrl);
                                             break;
 
-                case "BroadWorks Mobility" : this.broadWorksMobilityUrl = this.baseUrl + "broadworksmobility";
+                case 'BroadWorks Mobility' : this.broadWorksMobilityUrl = this.baseUrl + 'broadworksmobility';
                                              this.serviceRouter.setBroadWorksMobilityUrl(this.broadWorksMobilityUrl);
                                              break;
 
-                case "Simultaneous Ring Personal" : this.simultaneousRingUrl = this.baseUrl + "simultaneousringpersonal?";
+                case 'Simultaneous Ring Personal' : this.simultaneousRingUrl = this.baseUrl + 'simultaneousringpersonal?';
                                                     this.serviceRouter.setSimultaneousRingUrl(this.simultaneousRingUrl);
-                                                    break;            
+                                                    break;
 
-                case "Voice Messaging User": this.voiceMailUrl = this.baseUrl + "voicemessaging?";
-                    this.voicemailGreetingUrl = this.baseUrl + "voicemessaging/greetings?";
+                case 'Sequential Ring' : this.sequentialRingUrl = this.baseUrl + 'sequentialring';
+                                         this.serviceRouter.setSequentialRingUrl(this.sequentialRingUrl);
+                                         break;
+
+                case 'Voice Messaging User': this.voiceMailUrl = this.baseUrl + 'voicemessaging?';
+                    this.voicemailGreetingUrl = this.baseUrl + 'voicemessaging/greetings?';
                     this.serviceRouter.setVoicemailUrl(this.voiceMailUrl);
                     this.serviceRouter.setVoicemailGreetingUrl(this.voicemailGreetingUrl);
                     break;
-            }
 
+                case 'Call Center - Basic' : this.callCenterUrl = this.baseUrl + 'callcenter?';
+                    this.serviceRouter.setCallCenterUrl(this.callCenterUrl);
+                    break;
+
+                case 'Call Center - Standard': this.callCenterUrl = this.baseUrl + 'callcenter?';
+                    this.serviceRouter.setCallCenterUrl(this.callCenterUrl);
+                    break;
+
+                case 'Call Center - Premium': this.callCenterUrl = this.baseUrl + 'callcenter?';
+                    this.serviceRouter.setCallCenterUrl(this.callCenterUrl);
+
+                    break;
+
+                case 'BroadWorks Anywhere' : this.broadWorksAnywhereUrl = this.baseUrl + 'broadworksanywhere';
+                                             this.serviceRouter.setBroadWorksAnywhereUrl(this.broadWorksAnywhereUrl);
+                                             break;
+            }
         });
 
     }
@@ -151,6 +175,9 @@ export class ServiceRouteProvider {
     fetchAutomaticCallbackUrl() {
         return this.serviceRouter.getAutomaticCallbackUrl();
     }
+    fetchSequentialRingUrl() {
+        return this.serviceRouter.getSequentialRingUrl();
+    }
     fetchSimultaneousRingUrl() {
         return this.serviceRouter.getSimultaneousRingUrl();
     }
@@ -159,5 +186,16 @@ export class ServiceRouteProvider {
     }
     fetchVoicemailGreetingUrl() {
         return this.serviceRouter.getVoicemailGreetingUrl();
+    }
+    fetchCallCenterUrl() {
+        return this.serviceRouter.getCallCenterUrl();
+    }
+
+    fetchPersonalAssistantUrl() {
+        return this.serviceRouter.getPersonalAssistantUrl();
+    }
+
+    fetchBroadWorksAnywhereUrl() {
+        return this.serviceRouter.getBroadWorksAnywhereUrl();
     }
 }
