@@ -12,8 +12,9 @@ import { XSIMasterServicesList } from 'app/AppCommon/xsiMasterServicesList.servi
 import { MasterServiceListVisible } from 'app/AppCommon/masterServicesListVisible.service';
 @Injectable()
 
-/*This class prepares the services list that is to be displayed in the application*/
+/* This class prepares the master list of the services that is to be displayed in the app*/
 export class XSIServices {
+    broadWorksAnywhereVisible: any;
 
     private res: Response;
     private headers: Headers = new Headers();
@@ -21,137 +22,176 @@ export class XSIServices {
     private masterServicesList = [];
     private hiddenServicesList = [];
 
-    acrVisible: boolean = false;
-    cfVisible: boolean = false;
-    cfaVisible: boolean = false;
-    cfbVisible: boolean = false;
-    cfnaVisible: boolean = false;
-    cfnrVisible: boolean = false;
-    bcidVisible: boolean = false;
-    dndVisible: boolean = false;
-    roVisible: boolean = false;
-    cwVisible: boolean = false;
-    simultaneousRingVisible: boolean = false;
-    automaticCallbackVisible: boolean = false;
-    voicemailVisible: boolean = false;
-    bwMobilityVisible: boolean = false;
+    acrVisible = false;
+    cfVisible = false;
+    cfaVisible = false;
+    cfbVisible = false;
+    cfnaVisible = false;
+    cfnrVisible = false;
+    bcidVisible = false;
+    dndVisible = false;
+    roVisible = false;
+    cwVisible = false;
+    simultaneousRingVisible = false;
+    sequentialRingVisible = false;
+    automaticCallbackVisible = false;
+    voicemailVisible = false;
+    bwMobilityVisible = false;
+    personalAssistantVisible = false;
+    callCenterQueueVisible = false;
 
-    constructor(private masterServicesListVisible: MasterServiceListVisible, private xsiMasterServicesList: XSIMasterServicesList, private serviceRouteProvider: ServiceRouteProvider, private httpservices: HttpServices) { }
+    constructor (private masterServicesListVisible: MasterServiceListVisible, private xsiMasterServicesList: XSIMasterServicesList,
+      private serviceRouteProvider: ServiceRouteProvider, private httpservices: HttpServices) { }
 
     getXSIService(xsiUrl, postXSIGet) {
         this.httpservices.httpGetRequest(xsiUrl)
             .subscribe((res) => {
                 this.xsiServicesList = JSON.stringify(res.json());
                 console.log('XSI Services List: ', this.xsiServicesList);
-                var userServicesList = JSON.parse(this.xsiServicesList);
+                const userServicesList = JSON.parse(this.xsiServicesList);
 
-                var keepGoing, element1, element2, count;
-                var heroAppServicesList = window['callSettingsHeroInput'];
-                if (userServicesList["Services"]["service"]) {
-                    if(userServicesList["Services"]["service"].constructor === Array) {
-                        userServicesList["Services"]["service"].forEach(element1 => {
+                const heroAppServicesList = window['callSettingsHeroInput'];
+                if (userServicesList['Services']['service']) {
+                    if (userServicesList['Services']['service'].constructor === Array) {
+                        userServicesList['Services']['service'].forEach(element1 => {
                         this.masterServicesList.push({
-                            "name": element1.name.$,
-                            "visible": true,
-                            "url": element1['uri']
-
-
+                            'name': element1.name.$,
+                            'visible': true,
                         });
 
                     });
-                    } else if(userServicesList["Services"]["service"].constructor === Object){
+                    } else if (userServicesList['Services']['service'].constructor === Object) {
                         this.masterServicesList.push({
-                            "name": userServicesList.Services.service.name.$,
-                            "visible": true,
-                            "url": userServicesList.Services.service.uri.$
-                        })
+
+                            'name': userServicesList.Services.service.name.$,
+                            'visible': true,
+                        });
                     }
-                    
+
                 }
 
-
+                let isCallCenterVisible = false;
                 if (heroAppServicesList.services) {
 
                     this.masterServicesList.forEach(element => {
 
                         heroAppServicesList.services.forEach(heroElement => {
 
-                            if (element.name == heroElement.name && heroElement.visible == false) {
+                            if (element.name === heroElement.name && heroElement.visible === false) {
 
                                 element.visible = false;
                             }
 
-                            if (element.name == heroElement.name) {
+                            if (element.name === heroElement.name) {
 
-                                console.log("Checking on : " + element.name + " , with boolean set to : " + element.visible);
+                                console.log('Checking on : ' + element.name + ' , with boolean set to : ' + element.visible);
                                 switch (element.name) {
-                                    case "Do Not Disturb":
+                                    case 'Do Not Disturb':
                                         this.dndVisible = element.visible;
                                         this.masterServicesListVisible.setDndVisible(element.visible);
-
                                         break;
 
-                                    case "Call Forwarding Always":
+                                    case 'Call Forwarding Always':
                                         this.cfaVisible = element.visible;
                                         this.masterServicesListVisible.setCfaVisible(element.visible);
                                         break;
 
-                                    case "Call Forwarding Busy":
+                                    case 'Call Forwarding Busy':
                                         this.cfbVisible = element.visible;
                                         this.masterServicesListVisible.setCfbVisible(element.visible);
                                         break;
 
-                                    case "Call Forwarding No Answer":
+                                    case 'Call Forwarding No Answer':
                                         this.cfnaVisible = element.visible;
                                         this.masterServicesListVisible.setCfnaVisible(element.visible);
                                         break;
 
-                                    case "Call Forwarding Not Reachable":
+                                    case 'Call Forwarding Not Reachable':
                                         this.cfnrVisible = element.visible;
                                         this.masterServicesListVisible.setCfnrVisible(element.visible);
                                         break;
 
-                                    case "Automatic Callback":
+                                    case 'Automatic Callback':
                                         this.automaticCallbackVisible = element.visible;
                                         this.masterServicesListVisible.setAutomaticCallbackVisible(element.visible);
                                         break;
 
-                                    case "Anonymous Call Rejection":
+                                    case 'Anonymous Call Rejection':
                                         this.acrVisible = element.visible;
                                         this.masterServicesListVisible.setAcrVisible(element.visible);
                                         break;
 
-                                    case "Calling Line ID Delivery Blocking":
+                                    case 'Calling Line ID Delivery Blocking':
                                         this.bcidVisible = element.visible;
                                         this.masterServicesListVisible.setBcidVisible(element.visible);
                                         break;
 
-                                    case "Remote Office":
+                                    case 'Remote Office':
                                         this.roVisible = element.visible;
                                         this.masterServicesListVisible.setRoVisible(element.visible);
                                         break;
 
-                                    case "Call Waiting":
+                                    case 'Call Waiting':
                                         this.cwVisible = element.visible;
                                         this.masterServicesListVisible.setCwVisible(element.visible);
                                         break;
 
-                                    case "Simultaneous Ring Personal":
+                                    case 'Sequential Ring':
+                                        this.sequentialRingVisible = element.visible;
+                                        this.masterServicesListVisible.setSequentialRingVisible(element.visible);
+                                        break;
+
+                                    case 'Simultaneous Ring Personal':
                                         this.simultaneousRingVisible = element.visible;
                                         this.masterServicesListVisible.setSimultaneousRingVisible(element.visible);
                                         break;
 
-                                    case "Voice Messaging User":
+                                    case 'Voice Messaging User':
                                         this.voicemailVisible = element.visible;
                                         this.masterServicesListVisible.setVoicemailVisible(element.visible);
                                         break;
 
-                                    case "BroadWorks Mobility":
+                                    case 'BroadWorks Mobility':
                                         this.bwMobilityVisible = element.visible;
                                         this.masterServicesListVisible.setBWMobilityVisible(element.visible);
+                                        break;
+
+                                    case 'Personal Assistant' :
+                                            this.personalAssistantVisible = element.visible;
+                                            this.masterServicesListVisible.setPAVisible(element.visible);
+                                        break;
+
+                                    case 'Call Center - Standard' :
+                                            this.callCenterQueueVisible = element.visible;
+                                            if(this.callCenterQueueVisible) {
+                                                isCallCenterVisible = true;
+                                            }
+                                            this.masterServicesListVisible.setcallCenterVisible(element.visible);
+                                            break;
+
+                                    case 'Call Center - Basic' :
+                                            this.callCenterQueueVisible = element.visible;
+                                            if(this.callCenterQueueVisible) {
+                                                isCallCenterVisible = true;
+                                            }
+                                            this.masterServicesListVisible.setcallCenterVisible(element.visible);
+                                            break;
+                                            
+                                    case 'Call Center - Premium' :
+                                            this.callCenterQueueVisible = element.visible;
+                                            if(this.callCenterQueueVisible) {
+                                                isCallCenterVisible = true;
+                                            }
+                                            this.masterServicesListVisible.setcallCenterVisible(element.visible);
+                                            break;
+
+                                    case 'BroadWorks Anywhere' :
+                                            this.broadWorksAnywhereVisible = element.visible;
+                                            this.masterServicesListVisible.setBroadWorksAnywhereVisible(element.visible);
+                                            break;
 
                                     default:
-                                        console.log("No Need to serve this service: ", element.name);
+                                        console.log('No Need to serve this service: ', element.name);
                                         break;
                                 }
                             }
@@ -160,6 +200,16 @@ export class XSIServices {
 
                     });
                 }
+
+                if(isCallCenterVisible) {
+                    this.masterServicesListVisible.setcallCenterVisible(true);
+                }
+
+                this.masterServicesList.forEach(element => {
+                    console.log('>>>>>>>>>>>>>>>>>>>Name: ', element.name)
+                });
+
+
 
                 this.xsiMasterServicesList.setXSIMasterServicesList(this.masterServicesList);
                 this.serviceRouteProvider.initializeServiceRoutes(this.masterServicesList);
@@ -226,6 +276,14 @@ export class XSIServices {
         this.masterServicesListVisible.setAcrVisible(isAcrVisible);
     }
 
+    fetchSequentialRingVisible() {
+        return this.masterServicesListVisible.getSequentialRingVisible();
+    }
+
+    setSequentialRingVisible(isSequentialRingVisible) {
+        this.masterServicesListVisible.setSequentialRingVisible(isSequentialRingVisible);
+    }
+
     fetchSimultaneousRingVisible() {
         return this.masterServicesListVisible.getSimultaneousRingVisible();
     }
@@ -268,17 +326,36 @@ export class XSIServices {
         this.masterServicesListVisible.setBWMobilityVisible(isBWMobilityVisible);
     }
 
+    fetchPAVisible() {
+        return this.masterServicesListVisible.getPAVisible();
+    }
+
+    fetchCallcenterQueueVisible() {
+        return this.masterServicesListVisible.getCallCenterVisible();
+    }
+    setCallcenterQueueVisible(isCallCenterVisible) {
+        this.masterServicesListVisible.setcallCenterVisible(isCallCenterVisible);
+    }
+
+    fetchBroadWorksAnywhereVisible() {
+        return this.masterServicesListVisible.getBroadWorksAnywhereVisible();
+    }
+
+    setBroadWorksAnywhereVisible(isBroadWorksAnywhereVisible) {
+        this.masterServicesListVisible.setBroadWorksAnywhereVisible(isBroadWorksAnywhereVisible);
+    }
+
     generateLoginTokenAndUse(postGeneratedLoginToken) {
 
-        var loginTokenUrl = window['xsiActionsBaseURL'] + '/v2.0/user/' + window['callSettingsHeroInput'].userId + '/profile/LoginToken';
+        const loginTokenUrl = window['xsiActionsBaseURL'] + '/v2.0/user/' + window['callSettingsHeroInput'].userId + '/profile/LoginToken';
 
-        console.log("generateLoginTokenAndUse invoked");
+        console.log('generateLoginTokenAndUse invoked');
 
         this.httpservices.loginTokenPostRequest(loginTokenUrl).subscribe((res) => {
 
-            var responseJSON = JSON.parse(res.text());
-            var generatedToken = responseJSON.LoginToken.token.$;
-            console.log("Development or QA mode Generated Login token : " + generatedToken);
+            const responseJSON = JSON.parse(res.text());
+            const generatedToken = responseJSON.LoginToken.token.$;
+            console.log('Development or QA mode Generated Login token : ' + generatedToken);
             window['callSettingsHeroInput'].authorization = window['callSettingsHeroInput'].authorization + generatedToken;
             postGeneratedLoginToken(null);
 
