@@ -2,8 +2,10 @@
 
 import { Component, OnInit } from '@angular/core';
 
+import {IncomingComponent} from 'app/IncomingCalls/incoming.component';
 import { SimultaneousRingService } from 'app/IncomingCalls/SimultaneousRing/simultaneousRing.service';
 import { ServiceRouteProvider } from 'app/AppCommon/serviceRouteProvider.service';
+import { XSIServices } from 'app/AppCommon/xsiServiceList.service';
 import { SimultaneousRingServiceInput } from 'app/IncomingCalls/SimultaneousRing/simultaneousRingServiceInput.service';
 import { SimRingArray } from 'app/IncomingCalls/SimultaneousRing/simultaneousRingServiceInput.service';
 import { CriteriaArray } from 'app/IncomingCalls/SimultaneousRing/simultaneousRingServiceInput.service';
@@ -14,56 +16,67 @@ import { Util } from 'app/AppCommon/util';
   providers: [SimultaneousRingService]
 })
 
-export class SimultaneousRingComponent {
+export class SimultaneousRingComponent implements OnInit {
 
-  private simRingArray: SimRingArray[] = new Array();
-  private duplicateSimRingIndex: number = 100;  //Initially should not point any object
-  private criteriaArray: CriteriaArray[] = new Array();
 
-  private isWhenToRingExpanded: boolean = false;
-  private isSimultaneousRingActive: boolean = false;
-  private isRingNumbersChecked: boolean = false;
-  private isDoNotRingWhenOnCallChecked: boolean = false;
-  private phoneNumberMaxLength: number;
+  private duplicateSimRingIndex = 100;  // initially should not point any object
+  private isSimultaneousRingActive = false;
+  private RING_NUMBERS_UPDATE_FLAG = 1; // Setting the value for the update choice
+  private CRITERIA_UPDATE_FLAG = 2;    // Setting the value for the update choice
+  private SIM_RING_DEFAULT_UPDATE_FLAG = 0; // Setting the value for default update choice
+  private simultaneousRingVisible: boolean;
+  simRingUpdateError: String = '';
+  simRingRetrievingError;
+  isSimRingUpdateInProgress = false;
+  isWhenToRingExpanded = false;
+  isDoNotRingWhenOnCallUpdateError = '';
+  isDoNotRingWhenOnCallUpdateInprogress = false;
+  isRingNumbersChecked = false;
+  isDoNotRingWhenOnCallChecked = false;
+  phoneNumberMaxLength: number;
+  isPreviousEnterUpdate = false;
+  customizedTextJson = window['customizedTexts'];
+  simRingArray: SimRingArray[] = new Array();
+  criteriaArray: CriteriaArray[] = new Array();
 
-  private RING_NUMBERS_UPDATE_FLAG: number = 1; //Setting the value for the update choice
-  private CRITERIA_UPDATE_FLAG: number = 2;    //Setting the value for the update choice
-  private SIM_RING_DEFAULT_UPDATE_FLAG: number = 0; //Setting the value for default update choice
-  private simRingUpdateError: String = "";
-  private isSimRingUpdateInProgress: boolean = false;
-  private isDoNotRingWhenOnCallUpdateError: string = "";
-
-  private isCriteriaUpdateError: string = "";
-  private isDoNotRingWhenOnCallUpdateInprogress: boolean = false;
-
-  private customizedTextJson = window['customizedTexts'];
-
-  constructor(private simultaneousRingService: SimultaneousRingService, private serviceRouteprovider: ServiceRouteProvider, private simultaneousRingServiceInput: SimultaneousRingServiceInput,
-    private util: Util) {
+  constructor(private simultaneousRingService: SimultaneousRingService, private serviceRouteProvider: ServiceRouteProvider,
+    private simultaneousRingServiceInput: SimultaneousRingServiceInput,
+    private util: Util, private xsiServices: XSIServices) {
     this.phoneNumberMaxLength = this.util.INPUTMAXLENGTH;
   }
 
   ngOnInit() {
 
-    this.isDoNotRingWhenOnCallChecked = this.simultaneousRingService.fetchIsDoNotRingWhenOnCallChecked();
-    this.isSimultaneousRingActive = this.simultaneousRingService.fetchIsSimultaneousRingActive();
-    this.isRingNumbersChecked = this.isSimultaneousRingActive;
-    this.criteriaArray = this.simultaneousRingService.fetchCriteriaArray();
+    if(IncomingComponent.simRingExpandGet === true) {
+      this.isDoNotRingWhenOnCallChecked = this.simultaneousRingService.fetchIsDoNotRingWhenOnCallChecked();
+      this.isSimultaneousRingActive = this.simultaneousRingService.fetchIsSimultaneousRingActive();
+      this.isRingNumbersChecked = this.isSimultaneousRingActive;
+      this.criteriaArray = this.simultaneousRingService.fetchCriteriaArray();
+      this.simRingArray = this.simultaneousRingService.fetchSimRingArray();
+      IncomingComponent.simRingExpandGet = false;
+      this.addDefaultPhoneNumberEntry();
+    } else {
+      if (this.serviceRouteProvider.fetchSimultaneousRingUrl()) {
+        this.simultaneousRingVisible = this.xsiServices.fetchSimultaneousRingVisible();
+      }
+      if (this.simultaneousRingVisible) {
+        this.simRingRetrievingError = '';
+        this.simRingUpdateError = '';
+        this.simultaneousRingService.getSimultaneousRingService(this.serviceRouteProvider.fetchSimultaneousRingUrl(),
+                                        this.postSimultaneousRingGet.bind(this));
+      }
+    }
     if (!this.criteriaArray) {
       this.criteriaArray = new Array();
-    }
-
-
-    this.simRingArray = this.simultaneousRingService.fetchSimRingArray();
+    }    
     if (!this.simRingArray) {
       this.simRingArray = new Array();
     }
-    this.addDefaultPhoneNumberEntry();
   }
 
   private addDefaultPhoneNumberEntry() {
     if (!this.simRingArray.length || (this.simRingArray[this.simRingArray.length - 1].getNumber() && this.simRingArray.length <= 9)) {
-      let simrArray = new SimRingArray("", true);   //for new entry
+      let simrArray = new SimRingArray('', true);        // for new entry
       if (!this.simRingArray.length) {
         simrArray.setError(this.customizedTextJson.simultaneous_ring.phone_number_required);
       }
@@ -71,30 +84,27 @@ export class SimultaneousRingComponent {
     }
   }
 
+  private onPhonenumberChange(event: any) {
+    // this.clearErrorMessages();
 
-  private onPhoneNumberKeydown(event) {
-    if (event.keyCode == 8 || event.key == "Backspace" || event.keyCode == 46 || event.key == "Delete") {
-      let inputId = event.srcElement.id;
-      let index = parseInt(inputId.charAt(0));
-      this.clearFieldErrorMessage(index);
-    }
-  }
-
-  private onPhoneNumberChange(event: any) {
     let inputId = event.srcElement.id;
     let index = parseInt(inputId.charAt(0));
     this.clearFieldErrorMessage(index);
-    this.simRingArray[index].setIsValidNumber(true);  //updates the input line color as green
+    this.simRingArray[index].setIsValidNumber(true);      // updates the input border color as green
+    if (event.key !== 'Enter') {
+      this.isPreviousEnterUpdate = false;
+    }
   }
 
   clearFieldErrorMessage(index: number) {
-    this.simRingArray[index].setError("");
+    this.simRingArray[index].setError('');
   }
 
   clearErrorMessages() {
-    this.simRingUpdateError = "";
+
+    this.simRingUpdateError = '';
     for (let i = 0; i < this.simRingArray.length; i++) {
-      this.simRingArray[i].setError("");
+      this.simRingArray[i].setError('');
     }
   }
 
@@ -107,37 +117,51 @@ export class SimultaneousRingComponent {
   }
 
   simultaneousRingNumberEntered(event) {
+
+
     let self = this;
     setTimeout(function () {
+
       let inputId = event.srcElement.id;
       let input = event.srcElement.value;
+
       let index = parseInt(inputId.charAt(0));
       self.simRingArray[index].setIsValidNumber(false);
-      if (input) {
-        self.simRingArray[index].setNumber(input);
-        self.deleteTempDeletedSimRingArrayElements();
-      } else if (index == self.simRingArray.length - 1) {   //Should make last phonenumber  empty
-        self.simRingArray[self.simRingArray.length - 1].setNumber("");
-        if (index) {
-          self.simRingArray[self.simRingArray.length - 1].setError("");
-        } else {
-          self.setDefaultNoReqError();
+
+      if (!self.isPreviousEnterUpdate) {
+        if (input) {
+          self.simRingArray[index].setNumber(input);
+          self.deleteTempDeletedSimRingArrayElements();
+          // } else if (index != self.simRingArray.length - 1) {
+          //   self.simultaneousRingNumberDelete(index);
+        } else if (index === self.simRingArray.length - 1) { // Should make last phonenumber  empty
+          self.simRingArray[self.simRingArray.length - 1].setNumber('');
+          if (index) {
+            self.simRingArray[self.simRingArray.length - 1].setError('');
+          } else {
+            self.setDefaultNoReqError();
+          }
+        }
+        self.validateDuplicatePhoneNumbers();
+
+        let isValidateData: boolean = self.validatePhoneNumbers(input, index);
+        if (!input && index !== self.simRingArray.length - 1) {
+          self.simultaneousRingNumberDelete(index, isValidateData);
+        }
+        if (isValidateData) {
+          if (!self.simRingArray[0].getNumber()) {
+            self.simultaneousRingServiceInput.setIsSimultaneousRingActive(false);
+            self.isRingNumbersChecked = false;
+          }
+          self.isSimRingUpdateInProgress = true;
+          self.simultaneousRingService.sendSimultaneousRingPut(self.serviceRouteProvider.fetchSimultaneousRingUrl(),
+            self.updateSimultaneousRingServiceCb.bind(self), self.RING_NUMBERS_UPDATE_FLAG);
         }
       }
-      self.validateDuplicatePhoneNumbers();
-      let isValidateData: boolean = self.validatePhoneNumbers(input, index);
-      if (!input && index != self.simRingArray.length - 1) {
-        self.simultaneousRingNumberDelete(index, isValidateData);
-      }
-      if (isValidateData) {
-        if (!self.simRingArray[0].getNumber()) {
-          self.simultaneousRingServiceInput.setIsSimultaneousRingActive(false);
-          self.isRingNumbersChecked = false;
-        }
-        self.isSimRingUpdateInProgress = true;
-        self.simultaneousRingService.sendSimultaneousRingPut(self.serviceRouteprovider.fetchSimultaneousRingUrl(), self.updateSimultaneousRingServiceCb.bind(self), self.RING_NUMBERS_UPDATE_FLAG);
-      }
+
     }, 100);
+
+
   }
 
   private setDefaultNoReqError() {
@@ -145,17 +169,17 @@ export class SimultaneousRingComponent {
   }
 
   private validatePhoneNumbers(number: string, index: number) {
-    return this.isAllPhoneNumbersValid() && !this.isNumberAlreadyExist(number, index) && this.hasNoDuplicatePhoneNumbers();
+    return this.isAllPhoneNumbersValid() && !this.isPNumberAlreadyExist(number, index) && this.hasNoDuplicatePhoneNumbers();
   }
 
-  private isAllPhoneNumbersValid() {  //Validates all phone numbers
+  private isAllPhoneNumbersValid() {// Validates all phone numbers
     let flag = true;
     if (this.simRingArray) {
       let sArray: SimRingArray;
       for (let i = 0; i < this.simRingArray.length; i++) {
         sArray = this.simRingArray[i];
         if (sArray.getNumber()) {
-          if (!sArray.isPersisted() || (sArray.isPersisted() && sArray.getPersistedNumber() != sArray.getNumber())) {
+          if (!sArray.isPersisted() || (sArray.isPersisted() && sArray.getPersistedNumber() !== sArray.getNumber())) {
             if (!(this.util.isE164valid(sArray.getNumber()) || this.util.isValidSipUri(sArray.getNumber()))) {
               sArray.setError(this.customizedTextJson.remote_office.invalid_phone_number_msg);
               flag = false;
@@ -167,14 +191,14 @@ export class SimultaneousRingComponent {
     return flag;
   }
 
-  private isNumberAlreadyExist(number: string, index: number) {
+  private isPNumberAlreadyExist(pnumber: string, index: number) {
     let noAlreadyExists = false;
     if (this.simRingArray.length > 1) {
-      for (let count = 0; count < this.simRingArray.length; count++) {
-        if (index != count) {
-          if (!this.simRingArray[count].isTempDeleted() && number && number == this.simRingArray[count].getNumber()) {
-            this.setAsDuplicatePhNumEntry(index);
-            this.setAsDuplicatePhNumEntry(count);
+      for (let i = 0; i < this.simRingArray.length; i++) {
+        if (index !== i) {
+          if (!this.simRingArray[i].isTempDeleted() && pnumber && pnumber === this.simRingArray[i].getNumber()) {
+            this.setAsDuplicatePhNumEntry(index, pnumber);
+            this.setAsDuplicatePhNumEntry(i, pnumber);
             noAlreadyExists = true;
           }
         }
@@ -186,7 +210,7 @@ export class SimultaneousRingComponent {
   private hasNoDuplicatePhoneNumbers() {
     if (this.simRingArray.length > 1) {
       for (let i = 0; i < this.simRingArray.length; i++) {
-        if (this.simRingArray[i].getNumber() && this.isNumberAlreadyExist(this.simRingArray[i].getNumber(), i)) {
+        if (this.simRingArray[i].getNumber() && this.isPNumberAlreadyExist(this.simRingArray[i].getNumber(), i)) {
           return false;
         }
       }
@@ -197,14 +221,14 @@ export class SimultaneousRingComponent {
   private validateDuplicatePhoneNumbers() {
     if (this.simRingArray.length > 1) {
       for (let i = 0; i < this.simRingArray.length; i++) {
-        if (this.simRingArray[i].getNumber() && !this.isNumberAlreadyExist(this.simRingArray[i].getNumber(), i)) {
-          this.simRingArray[i].setError("");
+        if (this.simRingArray[i].getNumber() && !this.isPNumberAlreadyExist(this.simRingArray[i].getNumber(), i)) {
+          this.simRingArray[i].setError('');
         }
       }
     }
   }
 
-  private setAsDuplicatePhNumEntry(index: number) {
+  private setAsDuplicatePhNumEntry(index: number, pnumber: string) {
 
     this.simRingArray[index].setError(this.customizedTextJson.error.numberexist);
   }
@@ -212,28 +236,29 @@ export class SimultaneousRingComponent {
   updateSimultaneousRingServiceCb(res) {
     this.isSimRingUpdateInProgress = false;
     if (!res || !(res.status >= 200 && res.status < 400)) {
-      let errorStatus = "";
+      let errorStatus = '';
       if (res && (res.status === 0 || res.status)) {
         if (res.status === 0) {
           this.simRingUpdateError = this.customizedTextJson.error.networkerror;
-        }
-        else if (res.status >= 400) {
-          errorStatus = " " + res.status;
+        } else if (res.status >= 400) {
+         errorStatus = ' ' + res.status;
           this.simRingUpdateError = this.util.frameErrorMessage(this.customizedTextJson.error.updatefailed, errorStatus);
         }
       }
       this.isSimRingUpdateInProgress = false;
-    }
-    else {      //populate pojo with server details
+    } else {    // populate pojo with server details
+
       this.markAllPhoneNumbersAsPersisted();
       this.clearErrorMessages();
-      if (this.simRingArray[this.simRingArray.length - 1].getNumber()) {   //for new entry
+      this.isPreviousEnterUpdate = true;
+      if (this.simRingArray[this.simRingArray.length - 1].getNumber()) {// for new entry
         if (this.simRingArray.length <= 9) {
-          let simRingArray = new SimRingArray("", true);
+          let simRingArray = new SimRingArray('', true);
           this.simRingArray.push(simRingArray);
         }
       }
     }
+
     this.updateSimultaniousRingHeader(false);
   }
 
@@ -273,25 +298,38 @@ export class SimultaneousRingComponent {
       this.deleteTempDeletedSimRingArrayElements();
     }
 
-    if (this.simRingArray.length == 1 || !this.simRingArray[0].getNumber()) {
+    if (this.simRingArray.length === 1 || !this.simRingArray[0].getNumber()) {
       this.setDefaultNoReqError();
     }
   }
 
   private deleteTempDeletedSimRingArrayElements() {
-    for (let index = 0; index < this.simRingArray.length; index++) {
-      if (this.simRingArray[index].isTempDeleted()) {
-        this.simRingArray.splice(index, 1);
-        index--;
+    for (let i = 0; i < this.simRingArray.length; i++) {
+      if (this.simRingArray[i].isTempDeleted()) {
+        this.simRingArray.splice(i, 1);
+        i--;
       }
     }
   }
 
   doNotRingWhenOnCallActive(event) {
-    this.isDoNotRingWhenOnCallUpdateError = "";
+    this.isDoNotRingWhenOnCallUpdateError = '';
     this.isDoNotRingWhenOnCallUpdateInprogress = true;
     this.simultaneousRingServiceInput.setIsDoNotRingWhenOnCallChecked(event.checked);
-    this.simultaneousRingService.sendSimultaneousRingPut(this.serviceRouteprovider.fetchSimultaneousRingUrl(), this.postDoNotRingWhenOnCallPutCB.bind(this), this.SIM_RING_DEFAULT_UPDATE_FLAG);
+    
+    this.simultaneousRingService.sendSimultaneousRingPut(this.serviceRouteProvider.fetchSimultaneousRingUrl(),
+      this.postDoNotRingWhenOnCallPutCB.bind(this), this.SIM_RING_DEFAULT_UPDATE_FLAG);
+  }
+
+  doNotRingWhenOnCallKeyup(doNotRingWhenOnCallChecked) {
+
+    this.isDoNotRingWhenOnCallUpdateError = '';
+    this.isDoNotRingWhenOnCallUpdateInprogress = true;
+    this.isDoNotRingWhenOnCallChecked = !doNotRingWhenOnCallChecked;
+    this.simultaneousRingServiceInput.setIsDoNotRingWhenOnCallChecked(this.isDoNotRingWhenOnCallChecked);
+    
+    this.simultaneousRingService.sendSimultaneousRingPut(this.serviceRouteProvider.fetchSimultaneousRingUrl(),
+      this.postDoNotRingWhenOnCallPutCB.bind(this), this.SIM_RING_DEFAULT_UPDATE_FLAG);
   }
 
   postDoNotRingWhenOnCallPutCB(res) {
@@ -299,13 +337,12 @@ export class SimultaneousRingComponent {
     this.isDoNotRingWhenOnCallUpdateInprogress = false;
     if (!res || !(res.status >= 200 && res.status < 400)) {
       this.isDoNotRingWhenOnCallChecked = !this.isDoNotRingWhenOnCallChecked;
-      let errorStatus = "";
+      let errorStatus = '';
       if (res && (res.status === 0 || res.status)) {
         if (res.status === 0) {
           this.isDoNotRingWhenOnCallUpdateError = this.customizedTextJson.error.networkerror;
-        }
-        else {
-          errorStatus = " " + res.status;
+        } else {
+          errorStatus = ' ' + res.status;
           this.isDoNotRingWhenOnCallUpdateError = this.util.frameErrorMessage(this.customizedTextJson.error.updatefailed, errorStatus);
         }
       }
@@ -314,6 +351,8 @@ export class SimultaneousRingComponent {
     } else {
       this.simultaneousRingServiceInput.setIsDoNotRingWhenOnCallChecked(this.isDoNotRingWhenOnCallChecked);
     }
+
+    console.log('Do Not Ring When On Call is set to : ', this.isDoNotRingWhenOnCallChecked);
   }
 
   postSimultaneousRingPut(res) {
@@ -322,7 +361,7 @@ export class SimultaneousRingComponent {
 
       this.isRingNumbersChecked = !this.isRingNumbersChecked;
 
-      if (res && res.status) {
+      if (res && (res.status === 0 || res.status)) {
         if (res.status === 0) {
           this.simRingUpdateError = this.customizedTextJson.error.networkerror;
         } else if (res.status >= 400) {
@@ -333,6 +372,8 @@ export class SimultaneousRingComponent {
 
     this.restoreTempDeletedSimRingArrayNumbers();
     this.updateSimultaniousRingHeader(true);
+
+    console.log('Simultaneous ring is set to : ', this.simultaneousRingService.fetchIsSimultaneousRingActive());
   }
 
   private restoreTempDeletedSimRingArrayNumbers() {
@@ -344,35 +385,41 @@ export class SimultaneousRingComponent {
       }
     });
   }
-
+  
   postSimultaneousRingGet(simultaneousRingParsedJson) {
     if (simultaneousRingParsedJson) {
+      this.isDoNotRingWhenOnCallChecked = this.simultaneousRingService.fetchIsDoNotRingWhenOnCallChecked();
+      this.isSimultaneousRingActive = this.simultaneousRingService.fetchIsSimultaneousRingActive();
+      this.isRingNumbersChecked = this.isSimultaneousRingActive;
+      this.criteriaArray = this.simultaneousRingService.fetchCriteriaArray();
+      this.simRingArray = this.simultaneousRingService.fetchSimRingArray();
       this.isSimultaneousRingActive = this.simultaneousRingService.fetchIsSimultaneousRingActive();
       this.isDoNotRingWhenOnCallChecked = this.simultaneousRingService.fetchIsDoNotRingWhenOnCallChecked();
-
+      this.addDefaultPhoneNumberEntry();
     } else {
+      console.log('Some error occured in getting Simultaneous Ring.');
     }
   }
 
-// Removing the invalid numbers by validating...
-  private removeInvalidSimRingNumbers() {
+  private removeInvalidSRPhNumbers() {
     let sArray: SimRingArray;
-    for (let index = 0; index < this.simRingArray.length; index++) {
-      sArray = this.simRingArray[index];
+    for (let i = 0; i < this.simRingArray.length; i++) {
+      sArray = this.simRingArray[i];
       if (!sArray.isPersisted()) {
-        if (!sArray.getNumber() || (sArray.getNumber() && !(this.util.isE164valid(sArray.getNumber()) || this.util.isValidSipUri(sArray.getNumber())))) { // Validitiy checked...
-          this.simRingArray.splice(index, 1);
-          index--;
-        } else if (!sArray.getPersistedNumber()) {  // If not a persisting number, it is removed.
-          this.simRingArray.splice(index, 1);
-          index--;
+        if (!sArray.getNumber() || (sArray.getNumber() && !(this.util.isE164valid(sArray.getNumber())
+          || this.util.isValidSipUri(sArray.getNumber())))) {
+          this.simRingArray.splice(i, 1);
+          i--;
+        } else if (!sArray.getPersistedNumber()) {
+          this.simRingArray.splice(i, 1);
+          i--;
         }
       } else {
-        if (!(this.util.isE164valid(sArray.getNumber()) || this.util.isValidSipUri(sArray.getNumber()))) {  // Resetting the previously existing valid number which was changed to invalid...
+        if (!(this.util.isE164valid(sArray.getNumber()) || this.util.isValidSipUri(sArray.getNumber()))) {
           sArray.setNumber(sArray.getPersistedNumber());
         }
 
-        if (sArray.getNumber() != sArray.getPersistedNumber()) {  
+        if (sArray.getNumber() !== sArray.getPersistedNumber()) {
           sArray.setNumber(sArray.getPersistedNumber());
         }
 
@@ -384,12 +431,14 @@ export class SimultaneousRingComponent {
   }
 
   private removeDuplicatePhNumbers() {
-    for (let index1 = 0; index1 < this.simRingArray.length; index1++) {
-      for (let index2 = 0; index2 < this.simRingArray.length; index2++) {
-        if (index1 == index2) continue;
-        if (this.simRingArray[index1].getNumber() == this.simRingArray[index2].getNumber()) {
-          this.simRingArray.splice(index2, 1);
-          index1--;
+    for (let i = 0; i < this.simRingArray.length; i++) {
+      for (let j = 0; j < this.simRingArray.length; j++) {
+        if (i === j) {
+          continue;
+        }
+        if (this.simRingArray[i].getNumber() === this.simRingArray[j].getNumber()) {
+          this.simRingArray.splice(j, 1);
+          i--;
           break;
         }
       }
@@ -398,38 +447,54 @@ export class SimultaneousRingComponent {
 
 
   ringNumberSwitch() {
-    this.removeInvalidSimRingNumbers();
+
+    if(this.isRingNumbersChecked) {
+      this.isRingNumbersChecked = false;
+    } else {
+      this.isRingNumbersChecked = true;
+    }
+    
+    this.removeInvalidSRPhNumbers();
     this.removeDuplicatePhNumbers();
     this.clearErrorMessages();
-    if (this.simRingArray.length == 1 && !this.simRingArray[0].getNumber()) {
+    if (this.simRingArray.length === 1 && !this.simRingArray[0].getNumber()) {
       this.setDefaultNoReqError();
     }
 
     if (this.simultaneousRingServiceInput.getSimRingArray().length >= 2) {
       this.simultaneousRingServiceInput.setIsSimultaneousRingActive(this.isRingNumbersChecked);
-    }
-    else {
+    } else {
       this.simultaneousRingServiceInput.setIsSimultaneousRingActive(false);
     }
     if (!this.simultaneousRingService.fetchIsSimultaneousRingActive() || this.simRingArray.length > 1) {
       this.isSimRingUpdateInProgress = true;
-      this.simRingUpdateError = "";
-      this.simultaneousRingService.sendSimultaneousRingPut(this.serviceRouteprovider.fetchSimultaneousRingUrl(), this.postSimultaneousRingPut.bind(this), this.SIM_RING_DEFAULT_UPDATE_FLAG);
+      this.simRingUpdateError = '';
+
+      this.simultaneousRingService.sendSimultaneousRingPut(this.serviceRouteProvider.fetchSimultaneousRingUrl(),
+        this.postSimultaneousRingPut.bind(this), this.SIM_RING_DEFAULT_UPDATE_FLAG);
     }
+    
   }
 
   whenToRingChecked() {
-    if (this.isWhenToRingExpanded)
+    if (this.isWhenToRingExpanded) {
       this.isWhenToRingExpanded = false;
-    else
+    }else {
       this.isWhenToRingExpanded = true;
+    }
   }
 
   onCriteriaUpdate(criteria: CriteriaArray) {
     this.simultaneousRingServiceInput.setCriteriaArray(criteria);
+    if(criteria.getIsActive()) {
+      criteria.setIsActive(false);
+    } else {
+      criteria.setIsActive(true);
+    }
     criteria.setUpdateInprogress(true);
-    criteria.setErrorMsg("");
-    this.simultaneousRingService.sendSimultaneousRingPut(this.serviceRouteprovider.fetchSimultaneousRingUrl(), this.criteriaUpdateCB.bind(this, criteria), this.CRITERIA_UPDATE_FLAG);
+    criteria.setErrorMsg('');
+    this.simultaneousRingService.sendSimultaneousRingPut(this.serviceRouteProvider.fetchSimultaneousRingUrl(),
+      this.criteriaUpdateCB.bind(this, criteria), this.CRITERIA_UPDATE_FLAG);
 
   }
 
@@ -438,13 +503,12 @@ export class SimultaneousRingComponent {
 
     if (!res || !(res.status >= 200 && res.status < 400)) {
       criteria.setIsActive(!criteria.getIsActive());
-      let errorStatus = "";
+      let errorStatus = '';
       if (res && (res.status === 0 || res.status)) {
-        errorStatus = " " + res.status;
+        errorStatus = ' ' + res.status;
         if (res.status === 0) {
           criteria.setErrorMsg(this.customizedTextJson.error.networkerror);
-        }
-        else {
+        }else {
           criteria.setErrorMsg(this.util.frameErrorMessage(this.customizedTextJson.error.updatefailed, errorStatus));
           this.isSimRingUpdateInProgress = false;
         }
@@ -454,17 +518,28 @@ export class SimultaneousRingComponent {
   }
 
   updateAnswerConfirmationRequired(simRing) {
-    this.simultaneousRingServiceInput.setSimRingArray(simRing);
-    this.simRingUpdateError = "";
-    if (this.isAllPhoneNumbersValid() && this.hasNoDuplicatePhoneNumbers()) {
-      this.simultaneousRingService.sendSimultaneousRingPut(this.serviceRouteprovider.fetchSimultaneousRingUrl(), this.answerConfirmationRequiredCB.bind(this), this.RING_NUMBERS_UPDATE_FLAG);
-    }
+    let self = this;
+    if(simRing.isAnswerConfirmationRequired) {
+        simRing.isAnswerConfirmationRequired = false;
+      } else {
+        simRing.isAnswerConfirmationRequired = true;
+      }
+      
+    setTimeout(function () {
+      
+      self.simultaneousRingServiceInput.setSimRingArray(simRing);
+      self.simRingUpdateError = '';
+      if (self.isAllPhoneNumbersValid() && self.hasNoDuplicatePhoneNumbers()) {
+        self.simultaneousRingService.sendSimultaneousRingPut(self.serviceRouteProvider.fetchSimultaneousRingUrl(),
+          self.answerConfirmationRequiredCB.bind(self), self.RING_NUMBERS_UPDATE_FLAG);
+      }
+    }, 100);
   }
 
   answerConfirmationRequiredCB(res) {
     this.isSimRingUpdateInProgress = false;
     if (!res || !(res.status >= 200 && res.status < 400)) {
-      if (res && res.status) {
+      if (res && (res.status === 0 || res.status)) {
         if (res.status === 0) {
           this.simRingUpdateError = this.customizedTextJson.error.networkerror;
         } else if (res.status >= 400) {
@@ -476,3 +551,4 @@ export class SimultaneousRingComponent {
     }
   }
 }
+
