@@ -1,6 +1,6 @@
 /* Copyright © 2017 BroadSoft Inc. */
 
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { AutomaticCallbackService } from 'app/OutgoingCalls/automaticCallBackService.service';
 import { BlockCallerIdService } from 'app/OutgoingCalls/blockCallerIdService.service';
 import { AppComponent } from 'app/AppCommon/app.component';
@@ -15,51 +15,52 @@ import { Util } from 'app/AppCommon/util';
   providers: [AppComponent, AutomaticCallbackService, BlockCallerIdService]
 })
 
-export class outgoingComponent {
+export class OutgoingComponent implements OnInit {
 
+  private baseUrl: string = window['callSettingsHeroInput'].xsp + '/com.broadsoft.xsi-actions/v2.0/user/'
+                           + window['callSettingsHeroInput'].userId + '/services/';
+  private automaticCallbackUrl: string;
+  private blockCallerIdUrl: string;
+  private masterServicesList = [];
+  isBcidServiceRetrievingError = '';
+  isBcidServiceFetched = false;
+  isAcServiceRetrievingError = '';
+  isAcServiceFetched = false;
+  bcidVisible = false;
+  automaticCallbackVisible = false;
   customizedTextJson = window['customizedTexts'];
-
-  isBlockMyCalleridChecked: boolean = false;
-  isAutomaticCallbackChecked: boolean = false;
-  baseUrl: string = window['callSettingsHeroInput'].xsp + "/com.broadsoft.xsi-actions/v2.0/user/" + window['callSettingsHeroInput'].userId + "/services/";
-  automaticCallbackUrl: string;
-  blockCallerIdUrl: string;
-  masterServicesList = [];
-  isBcidServiceRetrievingError: string = "";
-  isBcidServiceFetched: boolean = false;
-  isAcServiceRetrievingError: string = "";
-  isAcServiceFetched: boolean = false;
-  bcidVisible: boolean = false;
-  automaticCallbackVisible: boolean = false;
+  isBlockMyCalleridChecked = false;
+  isAutomaticCallbackChecked = false;
   isBcidServiceUpdateError: string;
   isAcServiceUpdateError: string;
 
-  constructor(private xsiServices: XSIServices, private serviceRouteProvider: ServiceRouteProvider, private appComponent: AppComponent,
-    private blockCallerIdService: BlockCallerIdService, private automaticCallbackService: AutomaticCallbackService,
-    private util: Util) { }
+  constructor(private xsiServices: XSIServices, private serviceRouteProvider: ServiceRouteProvider,
+              private appComponent: AppComponent, private blockCallerIdService: BlockCallerIdService,
+              private automaticCallbackService: AutomaticCallbackService, private util: Util) { }
 
 
 
   ngOnInit() {
 
-    this.isBcidServiceUpdateError = "";
-    this.isAcServiceUpdateError = "";
+    this.isBcidServiceUpdateError = '';
+    this.isAcServiceUpdateError = '';
 
     if (this.serviceRouteProvider.fetchAutomaticCallbackUrl()) {
       this.automaticCallbackVisible = this.xsiServices.fetchAutomaticCallbackVisible();
     }
     if (this.serviceRouteProvider.fetchBlockCallerIdUrl()) {
-      this.bcidVisible = this.xsiServices.fetchBcidVisible();
+        this.bcidVisible = this.xsiServices.fetchBcidVisible();
     }
 
     /*Fetch active status of all services under outgoing calls from the server*/
     if (this.automaticCallbackVisible) {
-      this.isAcServiceRetrievingError = ""
-      this.automaticCallbackService.getAutomaticCallbackService(this.serviceRouteProvider.fetchAutomaticCallbackUrl(), this.postACBGet.bind(this));
+      this.isAcServiceRetrievingError = ''
+      this.automaticCallbackService.getAutomaticCallbackService(this.serviceRouteProvider.fetchAutomaticCallbackUrl(),
+                                                  this.postACBGet.bind(this));
     }
 
     if (this.bcidVisible) {
-      this.isBcidServiceRetrievingError = "";
+      this.isBcidServiceRetrievingError = '';
       this.blockCallerIdService.getBlockCallerIdService(this.serviceRouteProvider.fetchBlockCallerIdUrl(), this.postBCIDGet.bind(this));
     }
 
@@ -70,8 +71,10 @@ export class outgoingComponent {
 
     this.isAcServiceFetched = true;
     if (acbParsedJson) {
-      this.isAutomaticCallbackChecked = (acbParsedJson["AutomaticCallback"]["active"]["$"] == "true")
+      this.isAutomaticCallbackChecked = (acbParsedJson['AutomaticCallback']['active']['$'] === 'true')
+      console.log('isACBChecked: ', this.isAutomaticCallbackChecked)
     } else {
+      console.log('Some Error in ACB')
       this.isAcServiceRetrievingError = this.customizedTextJson.error.unabletofetch;
     }
   }
@@ -79,11 +82,12 @@ export class outgoingComponent {
   postACBPut(res) {
 
     this.isAcServiceFetched = true;
+
     if (!res || !(res.status >= 200 && res.status < 400)) {
       this.isAutomaticCallbackChecked = !this.isAutomaticCallbackChecked;
 
       if (res) {
-        if (res.status == 0) {
+        if (res.status === 0) {
           this.isAcServiceUpdateError = this.customizedTextJson.error.networkerror;
         } else {
           this.isAcServiceUpdateError = this.util.frameErrorMessage(this.customizedTextJson.error.updatefailed, res.status);
@@ -98,7 +102,8 @@ export class outgoingComponent {
     if (!bcidParsedJson) {
       this.isBcidServiceRetrievingError = this.customizedTextJson.error.unabletofetch;
     } else {
-      this.isBlockMyCalleridChecked = (bcidParsedJson["CallingLineIDDeliveryBlocking"]["active"]["$"] == "true");
+      this.isBlockMyCalleridChecked = (bcidParsedJson['CallingLineIDDeliveryBlocking']['active']['$'] === 'true')
+      console.log('isBlockMyCalleridChecked: ', this.isBlockMyCalleridChecked)
     }
 
   }
@@ -110,7 +115,7 @@ export class outgoingComponent {
     if (!res || !(res.status >= 200 && res.status < 400)) {
       this.isBlockMyCalleridChecked = !this.isBlockMyCalleridChecked;
       if (res) {
-        if (res.status == 0) {
+        if (res.status === 0) {
           this.isBcidServiceUpdateError = this.customizedTextJson.error.networkerror;
         } else {
           this.isBcidServiceUpdateError = this.util.frameErrorMessage(this.customizedTextJson.error.updatefailed, res.status);
@@ -122,30 +127,34 @@ export class outgoingComponent {
   automaticCallbackActive() {
 
     this.isAcServiceFetched = false;
-    this.isAcServiceUpdateError = "";
+    this.isAcServiceUpdateError = '';
     if (this.isAutomaticCallbackChecked) {
+      console.log('isAutomaticCallbackChecked = ', this.isAutomaticCallbackChecked);
       this.isAutomaticCallbackChecked = false;
-    }
-    else {
+    } else {
+      console.log('isAutomaticCallbackChecked = ', this.isAutomaticCallbackChecked);
       this.isAutomaticCallbackChecked = true;
     }
 
-    this.automaticCallbackService.putAutomaticCallbackService(this.serviceRouteProvider.fetchAutomaticCallbackUrl(), this.isAutomaticCallbackChecked, this.postACBPut.bind(this));
+    this.automaticCallbackService.putAutomaticCallbackService(this.serviceRouteProvider.fetchAutomaticCallbackUrl(),
+                                        this.isAutomaticCallbackChecked, this.postACBPut.bind(this));
 
   }
 
   blockCallerIdActive() {
 
     this.isBcidServiceFetched = false;
-    this.isBcidServiceUpdateError = "";
+    this.isBcidServiceUpdateError = '';
     if (this.isBlockMyCalleridChecked) {
+      console.log('isBlockMyCalleridChecked = ', this.isBlockMyCalleridChecked);
       this.isBlockMyCalleridChecked = false;
-    }
-    else {
+    } else {
+      console.log('isBlockMyCalleridChecked = ', this.isBlockMyCalleridChecked);
       this.isBlockMyCalleridChecked = true;
     }
 
-    this.blockCallerIdService.putBlockCallerIdService(this.serviceRouteProvider.fetchBlockCallerIdUrl(), this.isBlockMyCalleridChecked, this.postBCIDPut.bind(this));
+    this.blockCallerIdService.putBlockCallerIdService(this.serviceRouteProvider.fetchBlockCallerIdUrl(),
+                                              this.isBlockMyCalleridChecked, this.postBCIDPut.bind(this));
 
   }
 
